Refactor activation tests and consolidate test utilities
- Moved activation tests from `src/transformer/test_activations.c` to `src/test_activations.c` for better organization.
- Added comprehensive tests for SiLU, ReLU, and GELU activations, including scalar and in-place array tests.
- Introduced global test counters and utility macros for assertions across multiple test files.
- Updated all test files to utilize the new assertion macros and global counters.
- Removed redundant test utility definitions from individual test files.
- Ensured consistent test result reporting across all test files using `PRINT_TEST_RESULTS_AND_EXIT()`.
